JCache 1.1; Automatic-Module-Name; and other potential fixes
Now passes the TCK tests that were backwards incompatible with 1.0. The
OSGI test is enabled now that jsr107 includes the bindings.

Added a Jigsaw module names (fixes #211) for Java 9 modules. The work
towards a clean Java 9 library (no javax.annotations, no Unsafe) is
pending. This will released as version 3.0 when ready.

Stricter expiration checks for #212. Previously a stale read of the
current time was used where it was perceived as benign, to avoid
excessive (expensive) calls to the system clock. This might have
unintentionally resulted in an AsyncLoadingCache not expiring a
future properly.

Replaces the unnecessary NonReentrantLock with the JDK's rentrant one.
In #203 a deadlock occurs, though it is not reproducable and most likely
is due to a known (and scary) Linux kernel bug in older versions. The
improvements to state management also means the lock is extremely cheap
so any gains in the slimmer version are not detectable.